pyproject.toml
[project]
name = "credit-scoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.5",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argcomplete==3.6.2",
    "asttokens==3.0.1",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachecontrol==0.14.3",
    "cachetools==5.5.2",
    "category-encoders==2.9.0",
    "certifi==2025.7.14",
    "cffi==2.0.0",
    "charset-normalizer==3.4.2",
    "cleo==2.1.0",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "crashtest==0.4.1",
    "cryptography==45.0.7",
    "cycler==0.12.1",
    "databricks-sdk==0.65.0",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "distlib==0.4.0",
    "docker==7.1.0",
    "dulwich==0.22.8",
    "executing==2.2.1",
    "fastapi==0.116.1",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "findpython==0.6.3",
    "flask==3.1.2",
    "fonttools==4.59.2",
    "fsspec==2025.7.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-auth==2.40.3",
    "graphene==3.4.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "imbalanced-learn==0.14.0",
    "imblearn==0.0",
    "importlib-metadata==8.7.0",
    "inquirerpy==0.3.4",
    "installer==0.7.0",
    "ipykernel==7.1.0",
    "ipython==9.7.0",
    "ipython-pygments-lexers==1.1.1",
    "itsdangerous==2.2.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.2.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "keyring==25.6.0",
    "kiwisolver==1.4.9",
    "lightgbm>=4.6.0",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.2.1",
    "mlflow==3.3.2",
    "mlflow-skinny==3.3.2",
    "mlflow-tracing==3.3.2",
    "more-itertools==10.7.0",
    "msgpack==1.1.1",
    "nest-asyncio==1.6.0",
    "numpy==2.3.3",
    "opentelemetry-api==1.37.0",
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-semantic-conventions==0.58b0",
    "packaging==25.0",
    "pandas==2.3.2",
    "parso==0.8.5",
    "patsy==1.0.2",
    "pbs-installer==2025.7.23",
    "pexpect==4.9.0",
    "pfzy==0.3.4",
    "pillow==11.3.0",
    "pipx==1.7.1",
    "pkginfo==1.12.1.2",
    "platformdirs==4.3.8",
    "poetry==2.1.4",
    "poetry-core==2.1.3",
    "poetry-plugin-shell==1.0.1",
    "prompt-toolkit==3.0.51",
    "protobuf==6.32.0",
    "psutil==7.1.3",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pyproject-hooks==1.2.0",
    "pytest-cov>=7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.1.0",
    "rapidfuzz==3.13.0",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "scikit-learn==1.7.2",
    "scipy==1.16.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "starlette==0.47.2",
    "statsmodels==0.14.5",
    "threadpoolctl==3.6.0",
    "tomlkit==0.13.3",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "trove-classifiers==2025.5.9.12",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "userpath==1.9.2",
    "uv==0.9.17",
    "uvicorn==0.35.0",
    "virtualenv==20.32.0",
    "waitress==3.0.2",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=xml"
